Test: refactor to a lostash.root.dir convention
