+++ logstash-core/build.gradle
@@ -105,7 +105,7 @@ tasks.register("javaTests", Test) {
 tasks.register("rubyTests", Test) {
     inputs.files fileTree("${projectDir}/lib")
     inputs.files fileTree("${projectDir}/spec")
-    systemProperty 'logstash.core.root.dir', projectDir.absolutePath
+    systemProperty 'logstash.root.dir', projectDir.parent
     include '/org/logstash/RSpecTests.class'
     include '/org/logstash/config/ir/ConfigCompilerTest.class'
     include '/org/logstash/config/ir/CompiledPipelineTest.class'
@@ -150,8 +150,6 @@ idea {
         scopes.PROVIDED.plus += [project.configurations.provided]
     }
 }
-
-def customJRubyDir = project.hasProperty("custom.jruby.path") ? project.property("custom.jruby.path") : ""
 
 dependencies {
     compile files("../vendor/jruby/lib/jruby.jar") // jruby-core.jar
